Moved function to it's own endpoint
